refact(eventhandler): refactor event handler methods
- split eventhandler to separate add and delete handlers
