+++ cmd/ndm_daemonset/probe/addhandler.go
@@ -1,0 +1,218 @@
+/*
+Copyright 2020 The OpenEBS Authors
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package probe
+
+import (
+	"fmt"
+
+	"github.com/openebs/node-disk-manager/blockdevice"
+	apis "github.com/openebs/node-disk-manager/pkg/apis/openebs/v1alpha1"
+	"github.com/openebs/node-disk-manager/pkg/partition"
+
+	"k8s.io/apimachinery/pkg/api/errors"
+	"k8s.io/klog"
+)
+
+// addBlockDeviceToHierarchyCache adds the given block device to the hierarchy of devices.
+// returns true if the device already existed in the cache. Else returns false
+func (pe *ProbeEvent) addBlockDeviceToHierarchyCache(bd blockdevice.BlockDevice) bool {
+	var deviceAlreadyExistsInCache bool
+	// check if the device already exists in the cache
+	_, ok := pe.Controller.BDHierarchy[bd.DevPath]
+	if ok {
+		klog.V(4).Infof("device: %s already exists in cache, "+
+			"the event was likely generated by a partition table re-read", bd.DevPath)
+		deviceAlreadyExistsInCache = true
+	}
+	if !ok {
+		klog.V(4).Infof("device: %s does not exist in cache, "+
+			"the device is now connected to this node", bd.DevPath)
+		deviceAlreadyExistsInCache = false
+	}
+
+	// in either case, whether it existed or not, we will update with the latest BD into the cache
+	pe.Controller.BDHierarchy[bd.DevPath] = bd
+	return deviceAlreadyExistsInCache
+}
+
+// addBlockDevice processed when an add event is received for a device
+func (pe *ProbeEvent) addBlockDevice(bd blockdevice.BlockDevice, bdAPIList *apis.BlockDeviceList) error {
+
+	pe.addBlockDeviceToHierarchyCache(bd)
+
+	/*
+		Cases when an add event is generated
+		1. A new disk is added to the cluster to this node -  the disk is first time in this cluster
+		2. A new disk is added to this node -  the disk was already present in the cluster and it was moved to this node
+		3. A disk was detached and reconnected to this node
+		4. An add event due to partition table reread . This may cause events to be generated without the disk
+			being physically removed this node - (when a new partition is created on the device also, its the same case)
+	*/
+
+	// check if the disk can be uniquely identified. we try to generate the UUID for the device
+	klog.V(4).Infof("checking if device: %s can be uniquely identified", bd.DevPath)
+	uuid, ok := generateUUID(bd)
+	// if UUID cannot be generated create a GPT partition on the device
+	if !ok {
+		klog.V(4).Infof("device: %s cannot be uniquely identified", bd.DevPath)
+		if len(bd.DependentDevices.Partitions) > 0 ||
+			len(bd.DependentDevices.Holders) > 0 {
+			klog.V(4).Infof("device: %s has holders/partitions. %+v", bd.DevPath, bd.DependentDevices)
+		} else {
+			klog.Infof("starting to create partition on device: %s", bd.DevPath)
+			d := partition.Disk{
+				DevPath:          bd.DevPath,
+				DiskSize:         bd.Capacity.Storage,
+				LogicalBlockSize: uint64(bd.DeviceAttributes.LogicalBlockSize),
+			}
+			if err := d.CreateSinglePartition(); err != nil {
+				klog.Errorf("error creating partition for %s, %v", bd.DevPath, err)
+				return err
+			}
+			klog.Infof("created new partition in %s", bd.DevPath)
+			return nil
+		}
+	} else {
+		bd.UUID = uuid
+		klog.V(4).Infof("uuid: %s has been generated for device: %s", uuid, bd.DevPath)
+		bdAPI, err := pe.Controller.GetBlockDevice(uuid)
+
+		if errors.IsNotFound(err) {
+			klog.V(4).Infof("device: %s, uuid: %s not found in etcd", bd.DevPath, uuid)
+			/*
+				Cases when the BlockDevice is not found in etcd
+				1. The device is appearing in this cluster for the first time
+				2. The device had partitions and BlockDevice was not created
+			*/
+
+			if bd.DeviceAttributes.DeviceType == blockdevice.BlockDeviceTypePartition {
+				klog.V(4).Infof("device: %s is partition", bd.DevPath)
+				klog.V(4).Info("checking if device has a parent")
+				// check if device has a parent that is claimed
+				parentBD, ok := pe.Controller.BDHierarchy[bd.DependentDevices.Parent]
+				if !ok {
+					klog.V(4).Infof("unable to find parent device for device: %s", bd.DevPath)
+					return fmt.Errorf("cannot get parent device for device: %s", bd.DevPath)
+				}
+
+				klog.V(4).Infof("parent device: %s found for device: %s", parentBD.DevPath, bd.DevPath)
+				klog.V(4).Infof("checking if parent device can be uniquely identified")
+				parentUUID, parentOK := generateUUID(parentBD)
+				if !parentOK {
+					klog.V(4).Infof("unable to generate UUID for parent device, may be a device without WWN")
+					// cannot generate UUID for parent, may be a device without WWN
+					// used the new algorithm to create partitions
+					return pe.createBlockDeviceResourceIfNoHolders(bd, bdAPIList)
+				}
+
+				klog.V(4).Infof("uuid: %s generated for parent device: %s", parentUUID, parentBD.DevPath)
+
+				parentBDAPI, err := pe.Controller.GetBlockDevice(parentUUID)
+
+				if errors.IsNotFound(err) {
+					// parent not present in etcd, may be device without wwn or had partitions/holders
+					klog.V(4).Infof("parent device: %s, uuid: %s not found in etcd", parentBD.DevPath, parentUUID)
+					return pe.createBlockDeviceResourceIfNoHolders(bd, bdAPIList)
+				}
+
+				if err != nil {
+					klog.Error(err)
+					return err
+					// get call failed
+				}
+
+				if parentBDAPI.Status.ClaimState != apis.BlockDeviceUnclaimed {
+					// device is in use, and the consumer is doing something
+					// do nothing
+					klog.V(4).Infof("parent device: %s is in use, device: %s can be ignored", parentBD.DevPath, bd.DevPath)
+					return nil
+				} else {
+					// the consumer created some partitions on the disk.
+					// So the parent BD need to be deactivated and partition BD need to be created.
+					// 1. deactivate parent
+					// 2. create resource for partition
+
+					pe.Controller.DeactivateBlockDevice(*parentBDAPI)
+					deviceInfo := pe.Controller.NewDeviceInfoFromBlockDevice(&bd)
+					existingBlockDeviceResource := pe.Controller.GetExistingBlockDeviceResource(bdAPIList, deviceInfo.UUID)
+					err := pe.Controller.PushBlockDeviceResource(existingBlockDeviceResource, deviceInfo)
+					if err != nil {
+						klog.Error(err)
+						return err
+					}
+					return nil
+				}
+
+			}
+
+			if bd.DeviceAttributes.DeviceType != blockdevice.BlockDeviceTypePartition &&
+				len(bd.DependentDevices.Partitions) > 0 {
+				klog.V(4).Infof("device: %s has partitions: %+v", bd.DevPath, bd.DependentDevices.Partitions)
+				return nil
+			}
+
+			return pe.createBlockDeviceResourceIfNoHolders(bd, bdAPIList)
+		}
+
+		if err != nil {
+			klog.Errorf("querying etcd failed: %+v", err)
+			return err
+		}
+
+		if bdAPI.Status.ClaimState != apis.BlockDeviceUnclaimed {
+			klog.V(4).Infof("device: %s is in use. update the details of the blockdevice", bd.DevPath)
+			deviceInfo := pe.Controller.NewDeviceInfoFromBlockDevice(&bd)
+			err = pe.Controller.PushBlockDeviceResource(bdAPI, deviceInfo)
+			if err != nil {
+				klog.Errorf("updating block device resource failed: %+v", err)
+				return err
+			}
+			return nil
+		}
+
+		klog.V(4).Infof("creating resource for device: %s with uuid: %s", bd.DevPath, bd.UUID)
+		deviceInfo := pe.Controller.NewDeviceInfoFromBlockDevice(&bd)
+		existingBlockDeviceResource := pe.Controller.GetExistingBlockDeviceResource(bdAPIList, deviceInfo.UUID)
+		err = pe.Controller.PushBlockDeviceResource(existingBlockDeviceResource, deviceInfo)
+		if err != nil {
+			klog.Errorf("creation of resource failed: %+v", err)
+			return err
+		}
+		return nil
+	}
+	return nil
+}
+
+// createBlockDeviceResourceIfNoHolders creates/updates a blockdevice resource if it does not have any
+// holder devices
+func (pe *ProbeEvent) createBlockDeviceResourceIfNoHolders(bd blockdevice.BlockDevice, bdAPIList *apis.BlockDeviceList) error {
+	if len(bd.DependentDevices.Holders) > 0 {
+		klog.V(4).Infof("device: %s has holder devices: %+v", bd.DevPath, bd.DependentDevices.Holders)
+		klog.V(4).Infof("skip creating BlockDevice resource")
+		return nil
+	}
+
+	klog.V(4).Infof("creating block device resource for device: %s with uuid: %s", bd.DevPath, bd.UUID)
+	deviceInfo := pe.Controller.NewDeviceInfoFromBlockDevice(&bd)
+	existingBlockDeviceResource := pe.Controller.GetExistingBlockDeviceResource(bdAPIList, deviceInfo.UUID)
+	err := pe.Controller.PushBlockDeviceResource(existingBlockDeviceResource, deviceInfo)
+	if err != nil {
+		klog.Error(err)
+		return err
+	}
+	return nil
+}
