+++ cmd/ndm_daemonset/probe/eventhandler.go
@@ -17,16 +17,12 @@ limitations under the License.
 package probe
 
 import (
-	"fmt"
-
 	"github.com/openebs/node-disk-manager/blockdevice"
 	"github.com/openebs/node-disk-manager/cmd/ndm_daemonset/controller"
-	apis "github.com/openebs/node-disk-manager/pkg/apis/openebs/v1alpha1"
 	"github.com/openebs/node-disk-manager/pkg/features"
-	"github.com/openebs/node-disk-manager/pkg/partition"
 	libudevwrapper "github.com/openebs/node-disk-manager/pkg/udev"
 	"github.com/openebs/node-disk-manager/pkg/util"
-	"k8s.io/apimachinery/pkg/api/errors"
+
 	"k8s.io/klog"
 )
 
@@ -69,7 +65,7 @@ func (pe *ProbeEvent) addBlockDeviceEvent(msg controller.EventMessage) {
 		klog.Infof("Processed details for %s", device.DevPath)
 
 		if isGPTBasedUUIDEnabled {
-			err := pe.addBlockDevice(*device)
+			err := pe.addBlockDevice(*device, bdAPIList)
 			if err != nil {
 				isErrorDuringUpdate = true
 				klog.Error(err)
@@ -110,34 +106,18 @@ func (pe *ProbeEvent) deleteBlockDeviceEvent(msg controller.EventMessage) {
 	isGPTBasedUUIDEnabled := features.FeatureGates.IsEnabled(features.GPTBasedUUID)
 
 	for _, device := range msg.Devices {
-		// create the new UUID for removing the device
 		if isGPTBasedUUIDEnabled {
-			_, ok := pe.Controller.BDHierarchy[device.DevPath]
-			if !ok {
-				klog.Infof("Disk %s not in hierarchy", device.DevPath)
-				// not in hierarchy continue
+			_ = pe.deleteBlockDevice(*device, bdAPIList)
+		} else {
+			existingBlockDeviceResource := pe.Controller.GetExistingBlockDeviceResource(bdAPIList, device.UUID)
+			if existingBlockDeviceResource == nil {
+				// do nothing, may be the disk was filtered, or it was not created
+				isDeactivated = false
 				continue
 			}
-			// remove from the hierarchy
-			delete(pe.Controller.BDHierarchy, device.DevPath)
-
-			uuid, ok := generateUUID(*device)
-			// this can happen if the device didn't have any unique identifiers
-			if !ok {
-				klog.Info("could not recreate UUID while removing device")
-				continue
-			}
-			// use the new UUID for deactivating the blockdevice
-			device.UUID = uuid
+			pe.Controller.DeactivateBlockDevice(*existingBlockDeviceResource)
 		}
 
-		existingBlockDeviceResource := pe.Controller.GetExistingBlockDeviceResource(bdAPIList, device.UUID)
-		if existingBlockDeviceResource == nil {
-			// do nothing, may be the disk was filtered, or it was not created
-			isDeactivated = false
-			continue
-		}
-		pe.Controller.DeactivateBlockDevice(*existingBlockDeviceResource)
 	}
 
 	// rescan only if GPT based UUID is disabled.
@@ -174,7 +154,7 @@ func generateUUID(bd blockdevice.BlockDevice) (string, bool) {
 	// the partition and the unique data will be stored in sectors where consumers do not have access.
 
 	switch {
-	case bd.DeviceAttributes.DeviceType == libudevwrapper.UDEV_PARTITION:
+	case bd.DeviceAttributes.DeviceType == blockdevice.BlockDeviceTypePartition:
 		// The partition entry UUID is used when a partition (/dev/sda1) is processed. The partition UUID should be used
 		// if available, other than the partition table UUID, because multiple partitions can have the same partition table
 		// UUID, but each partition will have a different UUID.
@@ -196,194 +176,9 @@ func generateUUID(bd blockdevice.BlockDevice) (string, bool) {
 	}
 
 	if ok {
-		uuid = libudevwrapper.NDMBlockDevicePrefix + util.Hash(uuidField)
+		uuid = blockdevice.BlockDevicePrefix + util.Hash(uuidField)
 		klog.Infof("generated uuid: %s for device: %s", uuid, bd.DevPath)
 	}
 
 	return uuid, ok
-}
-
-func (pe *ProbeEvent) addBlockDevice(bd blockdevice.BlockDevice) error {
-
-	bdAPIList, err := pe.Controller.ListBlockDeviceResource(true)
-	if err != nil {
-		klog.Error(err)
-		return err
-	}
-
-	// check if the device already exists in the cache
-	_, ok := pe.Controller.BDHierarchy[bd.DevPath]
-	if ok {
-		klog.V(4).Infof("device: %s already exists in cache, "+
-			"the event was likely generated by a partition table re-read", bd.DevPath)
-	}
-	if !ok {
-		klog.V(4).Infof("device: %s does not exist in cache, "+
-			"the device is now connected to this node", bd.DevPath)
-	}
-
-	// in either case, whether is existed or not, we will update with the latest BD into the cache
-	pe.Controller.BDHierarchy[bd.DevPath] = bd
-
-	/*
-		Cases when an add event is generated
-		1. A new disk is added to the cluster to this node -  the disk is first time in this cluster
-		2. A new disk is added to this node -  the disk was already present in the cluster and it was moved to this node
-		3. A disk was detached and reconnected to this node
-		4. An add event due to partition table reread . This may cause events to be generated without the disk
-			being physically removed this node - (when a new partition is created on the device also, its the same case)
-	*/
-
-	// check if the disk can be uniquely identified. we try to generate the UUID for the device
-	klog.V(4).Infof("checking if device: %s can be uniquely identified", bd.DevPath)
-	uuid, ok := generateUUID(bd)
-	// if UUID cannot be generated create a GPT partition on the device
-	if !ok {
-		klog.V(4).Infof("device: %s cannot be uniquely identified", bd.DevPath)
-		if len(bd.DependentDevices.Partitions) > 0 ||
-			len(bd.DependentDevices.Holders) > 0 {
-			klog.V(4).Infof("device: %s has holders/partitions. %+v", bd.DevPath, bd.DependentDevices)
-		} else {
-			klog.Infof("starting to create partition on device: %s", bd.DevPath)
-			d := partition.Disk{
-				DevPath:          bd.DevPath,
-				DiskSize:         bd.Capacity.Storage,
-				LogicalBlockSize: uint64(bd.DeviceAttributes.LogicalBlockSize),
-			}
-			if err := d.CreateSinglePartition(); err != nil {
-				klog.Errorf("error creating partition for %s, %v", bd.DevPath, err)
-				return err
-			}
-			klog.Infof("created new partition in %s", bd.DevPath)
-			return nil
-		}
-	} else {
-		bd.UUID = uuid
-		klog.V(4).Infof("uuid: %s has been generated for device: %s", uuid, bd.DevPath)
-		bdAPI, err := pe.Controller.GetBlockDevice(uuid)
-
-		if errors.IsNotFound(err) {
-			klog.V(4).Infof("device: %s, uuid: %s not found in etcd", bd.DevPath, uuid)
-			/*
-				Cases when the BlockDevice is not found in etcd
-				1. The device is appearing in this cluster for the first time
-				2. The device had partitions and BlockDevice was not created
-			*/
-
-			if bd.DeviceAttributes.DeviceType == libudevwrapper.UDEV_PARTITION {
-				klog.V(4).Infof("device: %s is partition", bd.DevPath)
-				klog.V(4).Info("checking if device has a parent")
-				// check if device has a parent that is claimed
-				parentBD, ok := pe.Controller.BDHierarchy[bd.DependentDevices.Parent]
-				if !ok {
-					klog.V(4).Infof("unable to find parent device for device: %s", bd.DevPath)
-					return fmt.Errorf("cannot get parent device for device: %s", bd.DevPath)
-				}
-
-				klog.V(4).Infof("parent device: %s found for device: %s", parentBD.DevPath, bd.DevPath)
-				klog.V(4).Infof("checking if parent device can be uniquely identified")
-				parentUUID, parentOK := generateUUID(parentBD)
-				if !parentOK {
-					klog.V(4).Infof("unable to generate UUID for parent device, may be a device without WWN")
-					// cannot generate UUID for parent, may be a device without WWN
-					// used the new algorithm to create partitions
-					return pe.createBlockDeviceResourceIfNoHolders(bd, bdAPIList)
-				}
-
-				klog.V(4).Infof("uuid: %s generated for parent device: %s", parentUUID, parentBD.DevPath)
-
-				parentBDAPI, err := pe.Controller.GetBlockDevice(parentUUID)
-
-				if errors.IsNotFound(err) {
-					// parent not present in etcd, may be device without wwn or had partitions/holders
-					klog.V(4).Infof("parent device: %s, uuid: %s not found in etcd", parentBD.DevPath, parentUUID)
-					return pe.createBlockDeviceResourceIfNoHolders(bd, bdAPIList)
-				}
-
-				if err != nil {
-					klog.Error(err)
-					return err
-					// get call failed
-				}
-
-				if parentBDAPI.Status.ClaimState != apis.BlockDeviceUnclaimed {
-					// device is in use, and the consumer is doing something
-					// do nothing
-					klog.V(4).Infof("parent device: %s is in use, device: %s can be ignored", parentBD.DevPath, bd.DevPath)
-					return nil
-				} else {
-					// the consumer created some partitions on the disk.
-					// So the parent BD need to be deactivated and partition BD need to be created.
-					// 1. deactivate parent
-					// 2. create resource for partition
-
-					pe.Controller.DeactivateBlockDevice(*parentBDAPI)
-					deviceInfo := pe.Controller.NewDeviceInfoFromBlockDevice(&bd)
-					existingBlockDeviceResource := pe.Controller.GetExistingBlockDeviceResource(bdAPIList, deviceInfo.UUID)
-					err := pe.Controller.PushBlockDeviceResource(existingBlockDeviceResource, deviceInfo)
-					if err != nil {
-						klog.Error(err)
-						return err
-					}
-					return nil
-				}
-
-			}
-
-			if bd.DeviceAttributes.DeviceType != libudevwrapper.UDEV_PARTITION &&
-				len(bd.DependentDevices.Partitions) > 0 {
-				klog.V(4).Infof("device: %s has partitions: %+v", bd.DevPath, bd.DependentDevices.Partitions)
-				return nil
-			}
-
-			return pe.createBlockDeviceResourceIfNoHolders(bd, bdAPIList)
-		}
-
-		if err != nil {
-			klog.Errorf("querying etcd failed: %+v", err)
-			return err
-		}
-
-		if bdAPI.Status.ClaimState != apis.BlockDeviceUnclaimed {
-			klog.V(4).Infof("device: %s is in use. update the details of the blockdevice", bd.DevPath)
-			deviceInfo := pe.Controller.NewDeviceInfoFromBlockDevice(&bd)
-			err = pe.Controller.PushBlockDeviceResource(bdAPI, deviceInfo)
-			if err != nil {
-				klog.Errorf("updating block device resource failed: %+v", err)
-				return err
-			}
-			return nil
-		}
-
-		klog.V(4).Infof("creating resource for device: %s with uuid: %s", bd.DevPath, bd.UUID)
-		deviceInfo := pe.Controller.NewDeviceInfoFromBlockDevice(&bd)
-		existingBlockDeviceResource := pe.Controller.GetExistingBlockDeviceResource(bdAPIList, deviceInfo.UUID)
-		err = pe.Controller.PushBlockDeviceResource(existingBlockDeviceResource, deviceInfo)
-		if err != nil {
-			klog.Errorf("creation of resource failed: %+v", err)
-			return err
-		}
-		return nil
-	}
-	return nil
-}
-
-// createBlockDeviceResourceIfNoHolders creates/updates a blockdevice resource if it does not have any
-// holder devices
-func (pe *ProbeEvent) createBlockDeviceResourceIfNoHolders(bd blockdevice.BlockDevice, bdAPIList *apis.BlockDeviceList) error {
-	if len(bd.DependentDevices.Holders) > 0 {
-		klog.V(4).Infof("device: %s has holder devices: %+v", bd.DevPath, bd.DependentDevices.Holders)
-		klog.V(4).Infof("skip creating BlockDevice resource")
-		return nil
-	}
-
-	klog.V(4).Infof("creating block device resource for device: %s with uuid: %s", bd.DevPath, bd.UUID)
-	deviceInfo := pe.Controller.NewDeviceInfoFromBlockDevice(&bd)
-	existingBlockDeviceResource := pe.Controller.GetExistingBlockDeviceResource(bdAPIList, deviceInfo.UUID)
-	err := pe.Controller.PushBlockDeviceResource(existingBlockDeviceResource, deviceInfo)
-	if err != nil {
-		klog.Error(err)
-		return err
-	}
-	return nil
 }
